# Conflicts:
#	app/Services/EventService.php
#	composer.json
#	composer.lock
#	config/telegraph.php
#	database/seeders/DatabaseSeeder.php
#	database/seeders/UserSeeder.php
#	routes/web.php

+++ app/Http/Controllers/Profile/BlogController.php
@@ -75,6 +75,8 @@ class BlogController extends Controller
      */
     public function destroy(Blog $blog): RedirectResponse
     {
+        $this->authorize('delete', $blog);
+
         $this->blogService->destroy($blog);
 
         session()->flash('message', 'The blog has been deleted.');
